+++ src/finlang/Transpiler/InterfaceImplementGenerator.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using System;
 using System.Text;
 
 namespace finlang.Transpiler;
@@ -8,20 +9,24 @@ public class InterfaceImplementGenerator
 {
     private readonly CFileGenerator visitor;
     private readonly C99ClsEnumInterface cls;
+    StyleSettings styleSettings;
+    protected string NL => styleSettings.newLine;
+    protected string Indent => styleSettings.indent;
 
-    public InterfaceImplementGenerator(C99ClsEnumInterface cls)
+    public InterfaceImplementGenerator(C99ClsEnumInterface cls, StyleSettings styleSettings)
     {
         if (!cls.IsClass)
             throw new InvalidOperationException("Object has to be an class for this code.");
 
-        visitor = new CFileGenerator(cls);
+        visitor = new CFileGenerator(cls, styleSettings);
         this.cls = cls;
+        this.styleSettings = styleSettings;
     }
 
     public void GenerateVtables()
     {
         visitor.UseCFile();
-        var sb = cls.cFile.mainCode;
+        var sb = cls.cFile.mainCodeSb;
 
         var directInterfaces = GetDirectInterfaces();
         foreach (var directInterface in directInterfaces)
@@ -29,21 +34,21 @@ public class InterfaceImplementGenerator
             cls.SetHasVTable();
             cls.cFile.AddFqnDependency(directInterface);
 
-            sb.AppendLine($"\n// virtual table implementation for {directInterface.Name}. Note that this is extern'd.");
+            sb.Append($"{NL}// virtual table implementation for {directInterface.Name}. Note that this is extern'd.{NL}");
             var vtableStructName = InterfaceGenerator.GetInterfaceVtableStructName(Namer.GetCName(directInterface));
             var vtableInstanceName = GetVtableInstanceName(vtableStructName);
 
             var decl = $"const {vtableStructName} {vtableInstanceName}";
 
-            cls.hFile.mainCode.AppendLine("\n// vtable is extern to allow const initializations");
-            cls.hFile.mainCode.AppendLine($"extern {decl};");
-            sb.AppendLine($"{decl} = {{");
+            cls.hFile.mainCodeSb.Append($"{NL}// vtable is extern to allow const initializations{NL}");
+            cls.hFile.mainCodeSb.Append($"extern {decl};{NL}");
+            sb.Append($"{decl} = {{{NL}");
             foreach (var interfaceMethod in InterfaceGenerator.GetAllInterfaceMethods(directInterface))
             {
                 IMethodSymbol impMethod = cls.GetMethods().Single(m => m.Name == interfaceMethod.Name);
                 var mDecl = (MethodDeclarationSyntax)impMethod.DeclaringSyntaxReferences.Single().GetSyntax();
 
-                sb.Append($"    .{impMethod.Name} = ");
+                sb.Append($"{Indent}.{impMethod.Name} = ");
 
                 // need to cast to use 'void*' self parameter. Ex: (bool (*)(void *))
                 sb.Append($"({Namer.GetCName(impMethod.ReturnType)} (*)");
@@ -52,9 +57,9 @@ public class InterfaceImplementGenerator
 
                 sb.Append(')');
 
-                sb.AppendLine($"{Namer.GetCName(impMethod)},");
+                sb.Append($"{Namer.GetCName(impMethod)},{NL}");
             }
-            sb.AppendLine("};\n");
+            sb.Append($"}};{NL}{NL}");
         }
     }
 
@@ -70,7 +75,7 @@ public class InterfaceImplementGenerator
 
     internal void GenerateInterfaceConversions()
     {
-        var sb = cls.hFile.mainCode;
+        var sb = cls.hFile.mainCodeSb;
         string myTypeName = cls.GetCName();
 
         var directInterfaces = GetDirectInterfaces();
@@ -104,7 +109,7 @@ public class InterfaceImplementGenerator
         var superMethods = InterfaceGenerator.GetAllInterfaceMethods(directInterface);
         string firstMethodName = superMethods.First().Name;
         string conversionFunctionName = InterfaceGenerator.GetConversionFunctionName(myTypeName, superTypeName);
-        sb.AppendLine($"\n// Up conversion from {myTypeName} to {superTypeName} interface");
-        sb.AppendLine($"#define {conversionFunctionName}(self_arg)    ({superTypeName}){{ .obj = self_arg, .obj_vtable = (const {superVtableTypeName}*)(&{vtableInstanceName}.{firstMethodName}) }}");
+        sb.Append($"{NL}// Up conversion from {myTypeName} to {superTypeName} interface{NL}");
+        sb.Append($"#define {conversionFunctionName}(self_arg)    ({superTypeName}){{ .obj = self_arg, .obj_vtable = (const {superVtableTypeName}*)(&{vtableInstanceName}.{firstMethodName}) }}{NL}");
     }
 }