+++ src/finlang/Transpiler/TranspilerOptions.cs
@@ -5,4 +5,6 @@ public class TranspilerOptions
     public bool OutputVersionInfo = true;
     public bool OutputTimestamp = true;
     public bool OutputChecksum = true;
+
+    public StyleSettings StyleSettings = new();
 }